requirements/dev.txt
# This file pulls in everything a developer needs. If it's a basic package
# needed to run the site, it belongs in requirements/prod.txt. If it's a
# package for developers (testing, docs, etc.), it goes in this file.

django-debug-toolbar==0.9.4
django-extensions
dnspython==1.12.0
iscpy==1.05

-r prod.txt

# from funfactory

# Documentation
Sphinx==1.0.7

# Testing
nose==1.0.0
mock==0.7.2
-e git://github.com/jbalogh/django-nose.git#egg=django_nose
-e git://github.com/jbalogh/test-utils.git#egg=test-utils
